+++ src/Vector2.h
@@ -7,10 +7,10 @@
 
 class Vector2 {
 public:
-	int32_t mX = 0;
-	int32_t mY = 0;
+	float mX = 0;
+	float mY = 0;
 
-	Vector2(int32_t ix = 0, int32_t iy = 0) : mX(ix), mY(iy) {}
+	Vector2(float ix = 0, float iy = 0) : mX(ix), mY(iy) {}
 
 	inline void flip() { mX = -mX; mY = -mY; }
